WIP on refactoring toolbar
